disk: sdhc: Switch to clock frequency from DTS
The "spi-max-frequency" property already exists, but was unused. This
now sets the SPI clock frequency to this value (limited to 24MHz) once
initialisation is complete.

Due to the nature of the SPI API, it is necessary to have two separate
configuration structures to switch clock speed as some SPI drivers only
compare pointers to detect changes.

Fixes: #32996, Fixes: #28092, #22906

Improves issue reported in #26513, but does not implement CMD6 to
switch to High Speed mode.

Signed-off-by: Rich Barlow <rich@bennellick.com>